Refactor server.js Relates #22
